pyproject.toml
[project]
name = "awscloudopsagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.11.0",
    "bedrock-agentcore<=0.1.5",
    "bedrock-agentcore-starter-toolkit==0.1.14",
    "boto3",
    "botocore>=1.40.45",
    "click>=8.3.0",
    "fastapi>=0.118.0",
    "h11>=0.16.0",
    "httpx>=0.28.1",
    "idna>=3.10",
    "jmespath>=1.0.1",
    "mcp>=1.16.0",
    "pydantic>=2.11.9",
    "pydantic-core>=2.33.2",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich",
    "s3transfer>=0.14.0",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "starlette>=0.48.0",
    "strands-agents",
    "strands-agents-tools",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.2",
    "urllib3>=2.5.0",
    "uv>=0.9.4",
    "uvicorn[standard]>=0.37.0",
]
